css styling and fixed errors with order creation

+++ Inventory-main/src/main/java/com/selrvk/inventory/Product.java
@@ -6,11 +6,11 @@ public class Product {
     //private byte[] img;
     private String name;
     private int stock;
-    private int srp;
-    private int buyingPrice;
+    private double srp;
+    private double buyingPrice;
     private String manufacturer;
 
-    public Product(int id, String name, int stock, int srp, int buyingPrice, String manufacturer){
+    public Product(int id, String name, int stock, double srp, double buyingPrice, String manufacturer){
 
         this.id = id;
         //this.img = img;
@@ -21,7 +21,7 @@ public class Product {
         this.manufacturer = manufacturer;
     }
 
-    public Product(String name, int stock, int srp, int buyingPrice, String manufacturer){
+    public Product(String name, int stock, double srp, double buyingPrice, String manufacturer){
 
         this.id = 0;
         //this.img = img;
@@ -37,14 +37,14 @@ public class Product {
     //public byte[] getImg(){ return img; }
     public String getName(){ return name; }
     public int getStock(){ return stock; }
-    public int getSrp(){ return srp; }
-    public int getBuyingPrice(){ return buyingPrice; }
+    public double getSrp(){ return srp; }
+    public double getBuyingPrice(){ return buyingPrice; }
     public String getManufacturer(){ return manufacturer; }
     public void setName(String name){ this.name = name; }
     //public void setImg(byte[] img){ this.img = img; }
     public void setStock(int stock){ this.stock = stock; }
-    public void setSrp(int srp){ this.srp = srp; }
-    public void setBuyingPrice(int buyingPrice){ this.buyingPrice = buyingPrice; }
+    public void setSrp(double srp){ this.srp = srp; }
+    public void setBuyingPrice(double buyingPrice){ this.buyingPrice = buyingPrice; }
     public void setManufacturer(String manufacturer){ this.manufacturer = manufacturer; }
 
 }
